FCREPO-1010 Tika extraction branch fcrepo-1010
This branch is published to get feedback.
It uses tika 0.10, tika 1.0 is expected in November 2011.
It is a big jar, 24mb, so PermGen space has to be raised,
and it doubles the size of fedoragsearch.war
The branch adds two functions to GenericOperationsImpl:
- getDatastreamFromTika: retrieves the text only
- getDatastreamFromTikaWithMetadata: retrieves metadata also
The branch comes with a test suite in gsearch.test.fgs24_1010, 
where the two functions are tested on both Lucene and Solr.
The tests have docx, doc, and pdf datastreams,
but potentially all the Tika formats are available,
since the branch uses AutoDetectParser in Tika.