FCREPO-1019 Exploration of complex GSearch use cases
This is in response to Jonathan Green's mail on 17 october 2011:
"Something that I think needs to be considered when moving forward
with gsearch is that the index may not always share a 1 to 1
relationship with objects in fedora. In a very atomistic content model
perhaps the solr document is actually composed of parts from many
related objects. These types of decisions are currently very hard to
make in XSLTs. While I think XSLTs have a place in transforming
metadata, there needs to be something more."
The test case committed here has both 
- an index with contents from an itql query to risearch, thus
combining or joining several Fedora objects, 
- and the new feature that there may be more than one index document
per Fedora object, the example here has an index document for each of
three datastreams.